# trip-builder/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trip Builder API",
        "description": "Simple trip builder api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "api server"
        }
    ],
    "paths": {
        "/airlines": {
            "get": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Get list of airlines",
                "description": "Returns list of airlines",
                "operationId": "getAirlinesList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Store airline",
                "description": "Create new airline",
                "operationId": "postAirline",
                "requestBody": {
                    "description": "airline object",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Airline"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/airlines/{airline}": {
            "get": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Get airline information",
                "description": "Returns airline data",
                "operationId": "getAirlineById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Airline id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Update airline",
                "description": "update an airline",
                "operationId": "putAirline",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Airline id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "airline object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Airline"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Delete airline",
                "description": "delete an airline",
                "operationId": "deleteAirline",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Airline id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/airports": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get list of airports",
                "description": "Returns list of airports",
                "operationId": "getAirportsList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Airports"
                ],
                "summary": "Store airport",
                "description": "Create new airport",
                "operationId": "postAirport",
                "requestBody": {
                    "description": "Airport object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Airport"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/airports/{airport}": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get airport information",
                "description": "Returns airport data",
                "operationId": "getAirportById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Airport id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Airports"
                ],
                "summary": "Update airport",
                "description": "update an airport",
                "operationId": "putAirport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Airport id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Airport object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Airport"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Airports"
                ],
                "summary": "Delete airport",
                "description": "delete an airport",
                "operationId": "deleteAirport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Airport id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Airline": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "code"
                        ],
                        "properties": {
                            "code": {},
                            "name": {}
                        },
                        "type": "object"
                    }
                ]
            },
            "Airport": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "code"
                        ],
                        "properties": {
                            "code": {},
                            "name": {},
                            "latitude": {},
                            "longitude": {},
                            "timezone": {},
                            "city_id": {}
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}